requirements.txt
# Google Agent Development Kit (ADK)
google-adk>=1.0.0

# Core dependencies
asyncio-mqtt>=0.16.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

# LLM and AI models
litellm>=1.0.0
google-generativeai>=0.8.0

# Additional tools and utilities
requests>=2.31.0
aiohttp>=3.9.0
python-dotenv>=1.0.0

# For development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0